+++ src/snipskit/hermes/decorators.py
@@ -12,8 +12,8 @@ Example:
 
 .. code-block:: python
 
-    from snipskit.apps import HermesSnipsApp
-    from snipskit.decorators.hermes import intent
+    from snipskit.hermes.apps import HermesSnipsApp
+    from snipskit.hermes.decorators import intent
 
     class SimpleSnipsApp(HermesSnipsApp):