+++ src/snipskit/mqtt/__init__.py
@@ -1,0 +1,3 @@
+"""This module contains classes to create components that communicate with
+Snips services using the MQTT protocol directly.
+"""